Создание авторизации и прав пользователей 01-03-21
