+++ app/models.py
@@ -2,6 +2,9 @@ from app import db
 from typing import List
 
 
+
+
+
 class Users(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     public_id = db.Column(db.Integer)
